go/test: use shell code to create mosaics
As mosaic library is not supposed to create mosaics, but we do need
to test those, we have a shell code in test/ to create those.
For example, btrfs creates a file, mounts it with the help of block
loopback device, and creates a filesystem on top of it.

This patch removes the go/test code to create mos files, and adds calls
to these shell functions -- prepare() and cleanup() -- to create real
mosaics.

While at it, use defer more to guarantee we'll close the opens etc.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>